adding BoilingWater.java

+++ src/Calculations.java
@@ -5,9 +5,9 @@ public class Calculations {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter a number: ");
-        int num1 = scanner.nextInt();
+        double num1 = scanner.nextDouble();
         System.out.println("Enter another number: ");
-        int num2 = scanner.nextInt();
+        double num2 = scanner.nextDouble();
                System.out.println(num1 + " + " + num2 + " = " + (num1 + num2));
         System.out.println(num1 + " - " + num2 + " = " + (num1 - num2));
         System.out.println(num1 + " * " + num2 + " = " + (num1 * num2));
